Move UnsupportedReason to the resolution object
