+++ supervisor/const.py
@@ -431,17 +431,3 @@ class HostFeature(str, Enum):
     REBOOT = "reboot"
     SERVICES = "services"
     SHUTDOWN = "shutdown"
-
-
-class UnsupportedReason(str, Enum):
-    """Reasons for unsupported status."""
-
-    CONTAINER = "container"
-    DBUS = "dbus"
-    DOCKER_CONFIGURATION = "docker_configuration"
-    DOCKER_VERSION = "docker_version"
-    LXC = "lxc"
-    NETWORK_MANAGER = "network_manager"
-    OS = "os"
-    PRIVILEGED = "privileged"
-    SYSTEMD = "systemd"
